Ligo plugin: Change how default-storage is inferred
We used to consider the first declaration in *.storageList.ext file to
be the default storage. This led to users finding themselves unable to
do anything useful in the first declaration. See
https://github.com/pinnacle-labs/taqueria/issues/1872 for more context

This commit changes how default_storage is inferred. Users have to
declare a variable named default_storage to make it a default value
during deployments